requirements.txt
autopep8==2.0.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
discum==1.4.1
filetype==1.2.0
idna==3.4
pycodestyle==2.9.1
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.10.1
schedule==1.1.0
six==1.16.0
tomli==2.0.1
ua-parser==0.16.1
urllib3==1.26.12
websocket-client==0.59.0
